examples/mongo_giveaway_bot/requirements.txt
discord.py==1.7.2
odmantic
dpytools>=0.17.4b